Get Weather by location
In the GetWeather Class,
Initialize an async method
that gets weather by
location. The method divides
the Visual Crossing API
URL into three so that it
can encode the location
(the search query) merge
them together as one fullUrl.

Fetch the response from the
API server and await it.
Await the result of calling
the json() method on the
response.

Log this result (weatherData)
to the console and return it.

In index.js, import the
GetWeather Class and
make a new instance of
it, with a location.