Handle errors before parsing JSON
Add an IF block that checks if
the status of the response is
not ok. Inside the IF block,
add more blocks that
handle status-specific logic.
This ensures that the appropriate
errors can be thrown for errors
apart from those that would
normally be caught by the
catch block (these would
normally be caught by the
catch block - no internet
connection,
DNS lookup failures and
malformed URLs).

These handles catching of
the errors that could occur
after a succesful HTTP request
and the promise resolves
successfully (the response's
ok property becomes
false - indicating the
status code is outside the
2xx range)

Change fullSearchUrl to
fullUrl on line 22 and update its
reference on line 23.