Handle weather data
Initialize a handleWeatherData()
method that simply checks if the
JSON returned by
getWeatherLByLocation exists and
if it does, returns an object
with only the data required for
the weather app.

Call handleWeatherData() in
initialize().

If it does not exist,it throws an
error. In proper projects, should
run a schema validation
to check if the properties from the
JSON are actually returned and in
the right format.