Start preparing for search results
Add a main-container element
in the index.html. This container
would contain all elements for
rendering the search results.

Select the mainContainer element
in the constructor in the
weather-module GetWeather class
so that other methods can access
it.

Initialize a
buildLeftSectionForSearchResults()
method that creates all the
elements for the left
section that will render
the search results, gives them
appropriate classes,
attributes and append them
to their respective parents.

Even though
buildLeftSectionForSearchResults()
is curently called in initialize(),
as development of the app
progress it won't be called
here. It will be called
by a method that will
populate these elements
with the respective weather
data.