Prepare for right section
Move the creation of the and
class adding of the
searchResultsContainer
to the constructor, so that
it can be accessed by methods
that need it.

Initialize a
buildRightSectionForSearchResults()
method that creates all the
elements for the right
section that will render
the search results, gives them
appropriate classes,
textContent and append them
to their respective parents.

Have this method return the
necessary elements that will
be populated by the search results
as an object.