Create a DOM Module.
Create a DOM
module to handle creating and
appending of elements in the
DOM so that the weather module
can handle fetching and
rendering logic.

In the DOM module, initialize and
export a BuildDOM class.
Move the selection of the
mainContainer, the creation
and class adding of the
searchResultsContainer from
the GetWeather class
to the constructor
of the BuildDOM class.

Import the BuildDOM
class into the weather
module. Create a new
instance of the
BuildDom class in the
renderSearchResults()
method. Save the
invocation of the
buildLeftSectionForSearchResults()
method to a constant so that
it can be used accordingly.