Prepare to reuse searchContainer
Initialize a buildSearchContainer()
method that takes in a parent
parameter. Move the creation of
the searchContainer, searchField
and searchButton from the
buildRightSectionForSearchResults()
to the
buildSearchContainer().

Append the searchContainer
to its parent(parameter)
and return the searchContainer.
This is so that the searchContainer
and its containing elements
can be reused anywhere it
is needed.

In
buildLeftSectionForSearchResults(),
call buildSearchContainer()
and pass the leftContainer
as an argument. From line 100,
append it to the
homeAndSearchContainer.