Build home page
Have buildSearchContainer()
return the searchField and
searchButton so that it
can be accessed by methods.

Initialize a buildHomePage()
that simply appends the
searchContainer to the
mainContainer (by passing
in mainContainer as an
argument to
buildSearchContainer()).

Return searchField and
searchButton from
buildHomePage() so that
it can be accessed by a
method that will handle
clicking on the
searchButton.