Refactor GetWeather
Refactor Getweather so that
its relevant methods can
take in a searchLocation
argument and work based
on the location.

First, remove the constructor
parameter and the correstponding
property.

Have prepareUrl()
take in a searchLocation
parameter and then
encode the searchlocation
instead of the location
that was defined in the
class's constructor.

Also have getWeatherByLocation()
take in a searchLocation
parameter so that this
parameter can be passed into
prepareUrl to get the full url
(fullUrl)
that would be fetched (response)

Have handleWeatherData() take in
a searchLocation parameter so that
searchLocation can be passed into
getWeatherByLocation (weatherData).

Also have renderSearchResults()
take in a searchLocation parameter
so that
searchLocation can be passed into
handleWeatherData (weatherData)

This setup ensures tha
these methods
can be called with the value
of the search field and the
weather data for the location
will be fetched, handled and
rendered.