Handle searchButton clicks
First select the mainContainer
in the constructor so that
methods can access it.

Initialize a
fetchHandleAndRenderWeatherData()
method that takes in
searchLocation parameter.
This method calls the
relevant methods that
fetch, handle and render
weather data and passes
each of them the
searchLocation parameter.

Initialize a
handleSearchButtonClicks()
method that adds a
click event listener on
the search button on
the homepage and on
click, it passes the
search field's value
into
fetchHandleAndRenderWeatherData()

In initialize(), call the
buildHomepage() method
of the BuildDOM class.
Call handleSearchButtonClicks()
and remove other methods.

This fetches, handles and
renders the weather data
of the location searched
for via the searchField
and searchButton.