Handle searchButton click on searchResults page
Call the handleSearchButtonClicks()
in the try and catch blocks of
renderSearchResults() so that
after the search results have
rendered, gets the searchField
and searchButton in the
search results page ready for
another search.

In the handleSearchButtonClicks(),
change the searchButton to
a let variable so that
it can be reassigned. Clone it
(The cloned element does not
carry over any attached event listeners)

Replaces the original searchButton
in the DOM with the cloned version.
This effectively removes any previously
attached event listeners from the
original button, since the cloned
element starts fresh. Update the
searchButton variable to refer
to the new clone.

Remove any previously attached
event listeners to prevent
duplicate event listeners.