Clear DOM before searchResults
Previously search results were
appended to the home page
search container and new
search results were appended
on searchResultsContainer,
next to old results.

Move the creation and class adding
of the search container to the
constructor so that it
can be accessed by other methods.
Update its references in the
buildSearchContainer().

In
buildLeftSectionForSearchResults(),
set the textContents of
searchContainer and
searchResultsContainer so that
they are cleared before any
form of rendering. This setup
effectively clears the
searchContainer from home page
and previous search results
before rendering search results.