Update renderSearchResults()
Previously, when the user
searches for weather data,
the DOM was being
painted even before
renderSearchResults() is
awaiting data from
handleWeatherData().

This was happening because
leftSectionElements and
rightSectionElements were
initialize before the
try catch block.

Move their initialization
inside the try catch block
to that the DOM gets painted
only if there is a
response (weatherData)