Add loading component
In the HTML, add a
loading-div, that will
be displayed from the time
the form is submitted to
the time the data is fetched.

In the CSS, style the div,
give it a display of none
and an appropriate
spinning animation.
Style the div's 'show'
class with a display of
block.

In the JS, select the
loadingDiv in the
constructor.
Add the 'show' class at
the top of
getWeatherByLocation()
so that the loading
animation is shown once
the method is called.

Add a finally block to
getWeatherByLocation()that
removes the 'show' class
that the loading animation
is always hidden after the
API call, regardless of
the outcome.