Dynamically update units
The default weather unit group
is for visualcrossing's API is
US - so temperatures will
be in Fahrenheit and wind speed
will be mph. Create properties
with initial values
in the constructor for the
weatherUnitGroup, and
temperatureSignUnit and the
windSpeedUnit. Create a
select field and two options
in the index.html and
select the select field
in the constructor.

The select field will be used
to toggle between two
weather unit groups.

Add these properties in their
respective positions in the
object that handleWeatherData()
returns, so that when the toggle
setup is complete, these units
can be updated dynamically.