Add unit group to search query
Refactor prepareUrl() to
take in the weatherUnitGroup
parameter. Pass in this
parameter into fullUrl where
it should be according to the
visualcrossing's API's docs.

In getWeatherByLocation(),
make fullUrl conditional by
calling prepareUrl() with
the second argument
for US and UK respectively.
Also have the temperatureUnitSign
and windSpeedUnit conditional.

A change listen event will
be added to the select form
and the repective
weatherUnitGroup can be
selected.