Toggle weathergroup after search
Initialize a
searchFieldValueForWeatherUnitToggle
property in the constructor.
This property will hold the value
of the search field - the last
location whose weather details
was seached for so that after a
search is done, changing the value
of the select will rerun the fetch
with this search field value.

In handleButtonClicks, on
the click of the searchButton
the value of the search field
is assigned to
searchFieldValueForWeatherUnitToggle
and the
fetchHandleAndRenderWeatherData()
run only if the value of the
searchField was not an empty
string. This is to make sure that
searchFieldValueForWeatherUnitToggle
is not an empty string so that an
empty string is not passed into
fetchHandleAndRenderWeatherData()
when the select value is changed.

In toggleWeatherUnit(), add
a condition that checks if
searchFieldValueForWeatherUnitToggle
is not empty string and if it is not,
pass it into
fetchHandleAndRenderWeatherData().

This ensures that after a search,
changing the select value after
a search has been made,
fetches the weather data of the
same location, with the new
weatherUnitGroup.