Build iconpack
Add a new directory and
add some svg icons. Import
these icons in the
dom module and
initialize the buildIconPack()
method. The method
creates the respective
image element and sets
its source to the
respective svg image.

The method returns
the image references.

These images will be used
to change the look of the
page based on the data.